Refactor: 타입 폴더명 수정
