+++ app/(route)/activity/register/_components/IntroduceImage/index.tsx
@@ -4,7 +4,7 @@ import type { Dispatch, SetStateAction } from 'react';
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 
-import type { Activity, ActivityEdit, EditDetail, SubImage } from '@/_types/activities/register';
+import type { Activity, ActivityEdit, EditDetail, SubImage } from '@/_types/activities/formTypes';
 
 import { usePostImage } from '@/_hooks/use-post-image';
 import useModalState from '@/_hooks/useModalState';
