+++ app/(route)/activity/register/_components/RegisterLayout/index.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation';
 import postActivity from '@/_apis/activities/postActivity';
 import { useMutation } from '@tanstack/react-query';
 
-import type { Activity } from '@/_types/activities/register';
+import type { Activity } from '@/_types/activities/formTypes';
 
 import useModalState from '@/_hooks/useModalState';
