+++ app/(route)/activity/register/_components/SchedulePicker/index.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react';
 import dayjs from 'dayjs';
 import Image from 'next/image';
 
-import type { Activity, ActivityEdit, EditDetail, Schedule } from '@/_types/activities/register';
+import type { Activity, ActivityEdit, EditDetail, Schedule } from '@/_types/activities/formTypes';
 
 import useWindowSize from '@/_hooks/useWindowSize';
