+++ app/_apis/activities/patchActivity.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import type { ActivityEdit } from '@/_types/activities/register';
+import type { ActivityEdit } from '@/_types/activities/formTypes';
 
 import axiosInstance from '@/_libs/axios';
